+++ packages/retry-plugin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @module-federation/retry-plugin
 
+## 2.8.1
+
+### Patch Changes
+
+- Updated dependencies [d901e2c]
+  - @module-federation/sdk@2.8.1
+  - @module-federation/runtime@2.8.1
+
 ## 2.8.0
 
 ### Patch Changes